[TASK]: Add property setter helper based on table and record uid
