+++ Classes/Repository/AbstractRepository.php
@@ -36,7 +36,7 @@ use LMS\Facade\{Extbase\QueryBuilder, Extbase\TypoScriptConfiguration, Extbase\E
  */
 abstract class AbstractRepository extends \TYPO3\CMS\Extbase\Persistence\Repository
 {
-    use ProvidesCRUDActions, StaticCreation, ExtensionHelper, Collectionable, CacheQuery;
+    use ProvidesCRUDActions, PropertyManagement, StaticCreation, ExtensionHelper, Collectionable, CacheQuery;
 
     /**
      * Sets the defined Storage PID that is set in the TypoScript area
@@ -64,8 +64,6 @@ abstract class AbstractRepository extends \TYPO3\CMS\Extbase\Persistence\Reposit
 
     /**
      * Retrieve pid for the repository if it's set
-     *
-     * @return int
      */
     public function getPid(): int
     {
@@ -73,7 +71,7 @@ abstract class AbstractRepository extends \TYPO3\CMS\Extbase\Persistence\Reposit
     }
 
     /**
-     * @return string
+     *
      */
     protected function getExtensionKey(): string
     {
@@ -81,43 +79,7 @@ abstract class AbstractRepository extends \TYPO3\CMS\Extbase\Persistence\Reposit
     }
 
     /**
-     * Returns an array with initialized properties for requested record
      *
-     * @psalm-suppress PossiblyInvalidMethodCall
-     *
-     * @param int $uid
-     * @param string $table
-     *
-     * @return array
-     */
-    public function findRaw(int $uid, string $table): array
-    {
-        $builder = QueryBuilder::getQueryBuilderFor($table)->select('*')->from($table);
-
-        $where = $builder->expr()->eq('uid', $uid);
-
-        return (array)$builder->where($where)->execute()->fetch();
-    }
-
-    /**
-     * Retrieve property from table/uid association
-     *
-     * @param int $uid
-     * @param string $table
-     * @param string $property
-     *
-     * @return mixed
-     */
-    public function findRawProperty(int $uid, string $table, string $property)
-    {
-        return collect($this->findRaw($uid, $table))->get($property);
-    }
-
-    /**
-     * @param int $uid
-     * @param array $props
-     *
-     * @return \LMS\Facade\Assist\Collection
      */
     public function findById(int $uid, array $props = []): Collection
     {
@@ -132,10 +94,7 @@ abstract class AbstractRepository extends \TYPO3\CMS\Extbase\Persistence\Reposit
     }
 
     /**
-     * @param array $uidList
-     * @param array $props
      *
-     * @return \LMS\Facade\Assist\Collection
      */
     public function findByIds(array $uidList, array $props = []): Collection
     {
@@ -147,8 +106,7 @@ abstract class AbstractRepository extends \TYPO3\CMS\Extbase\Persistence\Reposit
     }
 
     /**
-     * @param array $props
-     * @return \LMS\Facade\Assist\Collection
+     *
      */
     public function all(array $props = []): Collection
     {
